Cleanup timeseries components to share base svg rendering component.
