+++ public/components/ResultsTimeseriesView.vue
@@ -332,7 +332,7 @@ export default Vue.extend({
 		attachScalingHandlers() {
 			const dragstarted = (d, index, elem) => {
 				this.highlightPixelX = null;
-				$('.vertical-line').hide();
+				this.$line.hide();
 			};
 
 			const dragged = (d, index, elem) => {
@@ -377,6 +377,11 @@ export default Vue.extend({
 		},
 		attachTranslationHandlers() {
 
+			const dragstarted = (d, index, elem) => {
+				this.highlightPixelX = null;
+				this.$line.hide();
+			};
+
 			const dragged = (d, index, elem) => {
 
 				if (this.selectedMicroMin === null) {
@@ -419,6 +424,7 @@ export default Vue.extend({
 
 			this.svg.selectAll('.axis-selection-rect')
 				.call(d3.drag()
+					.on('start', dragstarted)
 					.on('drag', dragged)
 					.on('end', dragended));
 		},
@@ -507,13 +513,6 @@ svg.axis {
 	line-height: 32px;
 	height: 32px;
 	width: 48px;
-}
-.timeseries-chart-col {
-	float: left;
-	position: relative;
-	line-height: 32px;
-	height: 32px;
-	width: calc(100% - 276px);
 }
 .timeseries-chart-axis {
 	float: left;
