+++ public/components/SparklinePreview.vue
@@ -1,8 +1,11 @@
 <template>
-	<div class="sparkline-container" v-observe-visibility="visibilityChanged" v-bind:class="{'is-hidden': !isVisible}">
-		<svg v-if="isLoaded" ref="svg" class="line-chart"></svg>
+
+	<div class="sparkline-container" v-observe-visibility="visibilityChanged">
+		<sparkline-svg
+			:timeseries-extrema="timeseriesExtrema"
+			:timeseries="timeseries">
+		</sparkline-svg>
 		<i class="fa fa-plus zoom-sparkline-icon" @click.stop="onClick"></i>
-		<div v-if="!isLoaded" v-html="spinnerHTML"></div>
 		<b-modal id="sparkline-zoom-modal" :title="timeseriesId"
 			@hide="hideModal"
 			:visible="zoomSparkline"
@@ -10,36 +13,28 @@
 			<sparkline-chart :timeseries="timeseries" v-if="zoomSparkline"></sparkline-chart>
 		</b-modal>
 	</div>
+
 </template>
 
 <script lang="ts">
 
 import * as d3 from 'd3';
-import _ from 'lodash';
 import Vue from 'vue';
 import SparklineChart from '../components/SparklineChart';
+import SparklineSvg from '../components/SparklineSvg';
 import { Dictionary } from '../util/dict';
-import { circleSpinnerHTML } from '../util/spinner';
-import { getters as routeGetters } from '../store/route/module';
+import { TimeseriesExtrema } from '../store/dataset/index';
 import { getters as datasetGetters, actions as datasetActions } from '../store/dataset/module';
 
 export default Vue.extend({
 	name: 'sparkline-preview',
 
 	components: {
+		SparklineSvg,
 		SparklineChart
 	},
 
 	props: {
-		margin: {
-			type: Object as () => any,
-			default: () => ({
-				top: 8,
-				right: 16,
-				bottom: 8,
-				left: 16
-			})
-		},
 		dataset: String as () => string,
 		xCol: String as () => string,
 		yCol: String as () => string,
@@ -49,14 +44,8 @@ export default Vue.extend({
 	data() {
 		return {
 			zoomSparkline: false,
-			entry: null,
 			isVisible: false,
-			hasRendered: false,
 			hasRequested: false,
-			xAxisTitle: '',
-			yAxisTitle: '',
-			xScale: null,
-			yScale: null
 		};
 	},
 	computed: {
@@ -65,10 +54,7 @@ export default Vue.extend({
 			return timeseries[this.dataset];
 		},
 		isLoaded(): boolean {
-			if (!this.timeseriesForDataset) {
-				return false;
-			}
-			return !!this.timeseriesForDataset[this.timeseriesId];
+			return !!this.timeseries;
 		},
 		timeseries(): number[][] {
 			if (!this.timeseriesForDataset) {
@@ -76,22 +62,20 @@ export default Vue.extend({
 			}
 			return this.timeseriesForDataset[this.timeseriesId];
 		},
-		spinnerHTML(): string {
-			return circleSpinnerHTML();
-		},
-		svg(): d3.Selection<SVGElement, {}, HTMLElement, any> {
-			return  d3.select(this.$svg);
-		},
-		$svg(): any {
-			return this.$refs.svg as any;
-		},
-		width(): number {
-			const dims = this.$svg.getBoundingClientRect();
-			return dims.width - this.margin.left - this.margin.right;
-		},
-		height(): number {
-			const dims = this.$svg.getBoundingClientRect();
-			return dims.height - this.margin.top - this.margin.bottom;
+		timeseriesExtrema(): TimeseriesExtrema {
+			if (!this.timeseries) {
+				return null;
+			}
+			return {
+				x: {
+					min: d3.min(this.timeseries, d => d[0]),
+					max: d3.max(this.timeseries, d => d[0])
+				},
+				y: {
+					min: d3.min(this.timeseries, d => d[1]),
+					max: d3.max(this.timeseries, d => d[1])
+				}
+			};
 		}
 	},
 	methods: {
@@ -101,73 +85,12 @@ export default Vue.extend({
 				this.requestTimeseries();
 				return;
 			}
-			if (this.isVisible && this.hasRequested && !this.hasRendered) {
-				this.injectTimeseries();
-			}
 		},
 		onClick() {
 			this.zoomSparkline = true;
 		},
 		hideModal() {
 			this.zoomSparkline = false;
-		},
-		clearSVG() {
-			this.svg.selectAll('*').remove();
-		},
-		injectSparkline() {
-			if (!this.$svg) {
-				return;
-			}
-
-			const timeseries = this.timeseries;
-
-			this.xScale = d3.scalePoint()
-				.range([0, this.width]);
-			this.xScale.domain(timeseries.map(d => d[0]));
-
-			const min = d3.min(timeseries, d => d[1]);
-			const max = d3.max(timeseries, d => d[1]);
-
-			this.yScale = d3.scaleLinear()
-				.domain([min, max])
-				.range([this.height, 0]);
-
-			const line = d3.line()
-				.x(d => this.xScale(d[0]))
-				.y(d => this.yScale(d[1]))
-				.curve(d3.curveLinear);
-
-			const className = 'line-chart';
-			const g = this.svg.append('g')
-				.attr('transform', `translate(${this.margin.left}, ${this.margin.top})`)
-				.attr('class', className);
-
-			g.datum(this.timeseries);
-
-			g.append('path')
-				.attr('fill', 'none')
-				.attr('class', 'line')
-				.attr('d', line);
-		},
-		injectTimeseries() {
-			if (_.isEmpty(this.timeseries)) {
-				return;
-			}
-
-			if (this.width <= 0) {
-				console.warn('Invalid width for line chart');
-				return;
-			}
-
-			if (this.height <= 0) {
-				console.warn('Invalid height for line chart');
-				return;
-			}
-
-			this.clearSVG();
-			this.injectSparkline();
-
-			this.hasRendered = true;
 		},
 		requestTimeseries() {
 			this.hasRequested = true;
@@ -177,10 +100,6 @@ export default Vue.extend({
 				yColName: this.yCol,
 				timeseriesColName: this.timeseriesCol,
 				timeseriesID: this.timeseriesId
-			}).then(() => {
-				if (this.isVisible) {
-					this.injectTimeseries();
-				}
 			});
 		}
 	}
@@ -190,22 +109,6 @@ export default Vue.extend({
 
 <style>
 
-svg.line-chart {
-	position: relative;
-	max-height: 32px;
-	width: 100%;
-	border: 1px solid rgba(0,0,0,0);
-}
-
-svg.line-chart g {
-	stroke: #666;
-	stroke-width: 2px;
-}
-/*
-svg.line-chart:hover g {
-	stroke: #00c6e1;
-}
-*/
 .zoom-sparkline-icon {
 	position: absolute;
 	right: 4px;
@@ -216,6 +119,7 @@ svg.line-chart:hover g {
 
 .sparkline-container {
 	position: relative;
+	width: 100%;
 }
 
 .sparkline-container:hover .zoom-sparkline-icon {
@@ -236,7 +140,4 @@ svg.line-chart:hover g {
 	max-width: 50%;
 }
 
-.is-hidden {
-	visibility: hidden;
-}
 </style>
